# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sistema de Gestão de Projetos API",
    "description": "API para gerenciar o ciclo de vida de projetos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/notifications/me": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "Minhas Notificações",
        "description": "Retorna todas as notificações não lidas do usuário autenticado. Requer autenticação de qualquer usuário válido.",
        "operationId": "get_my_notifications_api_notifications_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationRead"
                  },
                  "type": "array",
                  "title": "Response Get My Notifications Api Notifications Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/notifications/{notification_id}/mark-as-read": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Marcar Notificação como Lida",
        "description": "Marca uma notificação específica como lida. Valida se a notificação pertence ao usuário autenticado. Requer autenticação de qualquer usuário válido.",
        "operationId": "mark_notification_as_read_api_notifications__notification_id__mark_as_read_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/audit-logs/": {
      "get": {
        "tags": [
          "Admin: Audit Logs"
        ],
        "summary": "Get Audit Logs",
        "description": "Endpoint para consulta de logs de auditoria (apenas para administradores).\n\nRetorna uma lista paginada de todos os logs de auditoria do sistema,\nincluindo informações sobre o usuário que realizou a ação, a ação executada,\ndetalhes contextuais e timestamp.",
        "operationId": "get_audit_logs_api_admin_audit_logs__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "Número da página",
              "default": 1,
              "title": "Page"
            },
            "description": "Número da página"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "exclusiveMinimum": 0,
              "description": "Tamanho da página",
              "default": 20,
              "title": "Size"
            },
            "description": "Tamanho da página"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AuditLogRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditLogRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identificador único do log de auditoria",
            "example": "550e8400-e29b-41d4-a716-446655440012"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserInProject"
              },
              {
                "type": "null"
              }
            ],
            "description": "Dados do usuário que executou a ação"
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Tipo de ação executada",
            "example": "PROJECT_CREATED"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Detalhes contextuais da ação",
            "example": {
              "client": "Empresa ABC",
              "project_name": "Implementação MS365"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Data e hora da execução da ação",
            "example": "2025-01-15T10:30:00Z"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action",
          "timestamp"
        ],
        "title": "AuditLogRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NotificationRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Identificador único da notificação",
            "example": "550e8400-e29b-41d4-a716-446655440011"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Conteúdo da mensagem de notificação",
            "example": "Projeto 'Implementação MS365' avançou para fase de Elaboração"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read",
            "description": "Indica se a notificação foi lida pelo usuário",
            "example": false
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link",
            "description": "Link relacionado à notificação",
            "example": "/projects/550e8400-e29b-41d4-a716-446655440003"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Data e hora de criação da notificação",
            "example": "2025-01-15T16:45:00Z"
          }
        },
        "type": "object",
        "required": [
          "id",
          "message",
          "is_read",
          "created_at"
        ],
        "title": "NotificationRead"
      },
      "PaginatedResponse_AuditLogRead_": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Número total de itens disponíveis na consulta",
            "example": 150
          },
          "page": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Page",
            "description": "Número da página atual (baseado em 1)",
            "example": 2
          },
          "size": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Size",
            "description": "Número de itens por página",
            "example": 20
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Número total de páginas disponíveis",
            "example": 8
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditLogRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Lista de itens para a página atual"
          }
        },
        "type": "object",
        "required": [
          "total",
          "page",
          "size",
          "pages",
          "items"
        ],
        "title": "PaginatedResponse[AuditLogRead]"
      },
...
      "UserInProject": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Identificador único do usuário",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Endereço de email do usuário",
            "example": "joao.silva@empresa.com"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email"
        ],
        "title": "UserInProject"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}